+++ CHANGELOG.md
@@ -4,6 +4,12 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/)
 
+## 2024-08-08
+
+### Added
+
+- The Item Quiz now displays a Top 20 leaderboard.
+
 ## 2024-08-07
 
 ### Added